+++ src/components/Tabs/examples/09-with-count.jsx
@@ -6,6 +6,7 @@ export default createClass({
 	render() {
 		return (
 			<div>
+				<h4>Static Width</h4>
 				<Tabs hasFullWidthTabs={false}>
 					<Tabs.Tab count={12}>
 						<Tabs.Title>One</Tabs.Title>
@@ -20,6 +21,26 @@ export default createClass({
 						Three content
 					</Tabs.Tab>
 					<Tabs.Tab count={3} isDisabled>
+						<Tabs.Title>Three</Tabs.Title>
+						Three content
+					</Tabs.Tab>
+				</Tabs>
+
+				<h4>Variable Width</h4>
+				<Tabs hasFullWidthTabs={false}>
+					<Tabs.Tab count={1231321} isVariableCountWidth={true}>
+						<Tabs.Title>One</Tabs.Title>
+						One content
+					</Tabs.Tab>
+					<Tabs.Tab count={6546541612} isVariableCountWidth={true}>
+						<Tabs.Title>Two</Tabs.Title>
+						Two content
+					</Tabs.Tab>
+					<Tabs.Tab count={7} isVariableCountWidth={true}>
+						<Tabs.Title>Three</Tabs.Title>
+						Three content
+					</Tabs.Tab>
+					<Tabs.Tab count={123} isDisabled isVariableCountWidth={true}>
 						<Tabs.Title>Three</Tabs.Title>
 						Three content
 					</Tabs.Tab>