+++ CHANGELOG.md
@@ -19,3 +19,4 @@
 - added `isInvisible` to `SingleSelect`. We do not use a button under the hood for this component, but this simulates an invisible `Button`, styling-wise.
 - removed `kindOf` from `ToolTip`
 - added `isLight` to `ToolTip`. The default now has dark styling, and this adds light styling.
+- added `isVariableCountWidth` to `Tabs`. Allows count badges to have fixed or variable widths.